Move image and page classes to packages. Refactor categories recipes endpoint.
