+++ app/src/main/java/pjwstk/receipemate/app/controller/category/recipe/CategoryRecipeController.java
@@ -4,7 +4,7 @@ import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import pjwstk.receipemate.app.view.PageView;
+import pjwstk.receipemate.app.view.page.PageView;
 import pjwstk.receipemate.app.viewrepository.category.recipe.CategoryRecipesPageViewRepository;
 
 @CrossOrigin(origins = "*", allowedHeaders = "*")
