+++ app/src/main/java/pjwstk/receipemate/app/repository/ingredient/IngredientRepository.java
@@ -1,4 +1,4 @@
-package pjwstk.receipemate.app.repository;
+package pjwstk.receipemate.app.repository.ingredient;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
