+++ app/src/main/java/pjwstk/receipemate/app/view/category/recipe/CategoryRecipesPageView.java
@@ -2,7 +2,7 @@ package pjwstk.receipemate.app.view.category.recipe;
 
 import lombok.Getter;
 import lombok.Setter;
-import pjwstk.receipemate.app.view.PageView;
+import pjwstk.receipemate.app.view.page.PageView;
 import pjwstk.receipemate.app.view.category.CategoryView;
 
 @Getter
