+++ app/src/main/java/pjwstk/receipemate/app/view/image/ImageView.java
@@ -1,4 +1,4 @@
-package pjwstk.receipemate.app.view;
+package pjwstk.receipemate.app.view.image;
 
 import lombok.Getter;
 import lombok.Setter;
