+++ app/src/main/java/pjwstk/receipemate/app/view/recipe/RecipeDetailedView.java
@@ -2,7 +2,7 @@ package pjwstk.receipemate.app.view.recipe;
 
 import lombok.Getter;
 import lombok.Setter;
-import pjwstk.receipemate.app.view.ImageView;
+import pjwstk.receipemate.app.view.image.ImageView;
 import pjwstk.receipemate.app.view.category.CategoryView;
 import pjwstk.receipemate.app.view.recipe.ingredient.RecipeIngredientView;
 import pjwstk.receipemate.app.view.recipe.step.RecipeStepView;
