+++ app/src/main/java/pjwstk/receipemate/app/view/recipe/RecipeView.java
@@ -2,7 +2,7 @@ package pjwstk.receipemate.app.view.recipe;
 
 import lombok.Getter;
 import lombok.Setter;
-import pjwstk.receipemate.app.view.ImageView;
+import pjwstk.receipemate.app.view.image.ImageView;
 import pjwstk.receipemate.app.view.category.CategoryView;
 
 import java.time.LocalDateTime;
