+++ app/src/main/java/pjwstk/receipemate/app/viewrepository/recipe/RecipeViewRepository.java
@@ -4,9 +4,9 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import pjwstk.receipemate.app.exception.NotFoundException;
-import pjwstk.receipemate.app.view.PageView;
+import pjwstk.receipemate.app.view.page.PageView;
 import pjwstk.receipemate.app.view.recipe.RecipeView;
-import pjwstk.receipemate.app.viewfactory.PageViewFactory;
+import pjwstk.receipemate.app.viewfactory.page.PageViewFactory;
 import pjwstk.receipemate.app.model.AverageRateRecipe;
 import pjwstk.receipemate.app.repository.recipe.AverageRateRecipeRepository;
 import pjwstk.receipemate.app.viewfactory.recipe.RecipeViewFactory;
